get the latest version

+++ src/Presentation/Nop.Web/Controllers/BoardsController.cs
@@ -4,10 +4,10 @@ using System.Linq;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Nop.Core;
-using Nop.Core.Domain.Customers;
 using Nop.Core.Domain.Forums;
 using Nop.Core.Domain.Security;
 using Nop.Core.Rss;
+using Nop.Services.Customers;
 using Nop.Services.Forums;
 using Nop.Services.Localization;
 using Nop.Web.Factories;
@@ -19,12 +19,14 @@ using Nop.Web.Models.Boards;
 namespace Nop.Web.Controllers
 {
     [HttpsRequirement(SslRequirement.No)]
+    [AutoValidateAntiforgeryToken]
     public partial class BoardsController : BasePublicController
     {
         #region Fields
 
         private readonly CaptchaSettings _captchaSettings;
         private readonly ForumSettings _forumSettings;
+        private readonly ICustomerService _customerService;
         private readonly IForumModelFactory _forumModelFactory;
         private readonly IForumService _forumService;
         private readonly ILocalizationService _localizationService;
@@ -38,6 +40,7 @@ namespace Nop.Web.Controllers
 
         public BoardsController(CaptchaSettings captchaSettings,
             ForumSettings forumSettings,
+            ICustomerService customerService,
             IForumModelFactory forumModelFactory,
             IForumService forumService,
             ILocalizationService localizationService,
@@ -47,6 +50,7 @@ namespace Nop.Web.Controllers
         {
             _captchaSettings = captchaSettings;
             _forumSettings = forumSettings;
+            _customerService = customerService;
             _forumModelFactory = forumModelFactory;
             _forumService = forumService;
             _localizationService = localizationService;
@@ -191,6 +195,7 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
+        [IgnoreAntiforgeryToken]
         public virtual IActionResult ForumWatch(int id)
         {
             var watchTopic = _localizationService.GetResource("Forum.WatchForum");
@@ -252,6 +257,7 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
+        [IgnoreAntiforgeryToken]
         public virtual IActionResult TopicWatch(int id)
         {
             var watchTopic = _localizationService.GetResource("Forum.WatchTopic");
@@ -304,8 +310,7 @@ namespace Nop.Web.Controllers
             return View(model);
         }
 
-        [HttpPost]
-        [PublicAntiForgery]
+        [HttpPost]        
         public virtual IActionResult TopicMove(TopicMoveModel model)
         {
             if (!_forumSettings.ForumsEnabled)
@@ -326,7 +331,6 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
         public virtual IActionResult TopicDelete(int id)
         {
             if (!_forumSettings.ForumsEnabled)
@@ -376,7 +380,6 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
         [ValidateCaptcha]
         public virtual IActionResult TopicCreate(EditForumTopicModel model, bool captchaValid)
         {
@@ -501,7 +504,6 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
         [ValidateCaptcha]
         public virtual IActionResult TopicEdit(EditForumTopicModel model, bool captchaValid)
         {
@@ -512,7 +514,8 @@ namespace Nop.Web.Controllers
 
             if (forumTopic == null)
                 return RedirectToRoute("Boards");
-            var forum = forumTopic.Forum;
+
+            var forum = _forumService.GetForumById(forumTopic.ForumId);
             if (forum == null)
                 return RedirectToRoute("Boards");
 
@@ -622,7 +625,6 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
         public virtual IActionResult PostDelete(int id)
         {
             if (!_forumSettings.ForumsEnabled)
@@ -632,26 +634,28 @@ namespace Nop.Web.Controllers
                 });
 
             var forumPost = _forumService.GetPostById(id);
-            if (forumPost != null)
+
+            //TODO: ???? get topic one more time because it can be deleted (first or only post deleted)
+            var forumTopic = _forumService.GetTopicById(forumPost.TopicId);
+
+            if (forumPost != null && forumTopic != null)
             {
                 if (!_forumService.IsCustomerAllowedToDeletePost(_workContext.CurrentCustomer, forumPost))
                     return Challenge();
 
-                var forumTopic = forumPost.ForumTopic;
-                var forumId = forumTopic.Forum.Id;
-                var forumSlug = _forumService.GetForumSeName(forumTopic.Forum);
+                var forum = _forumService.GetForumById(forumTopic.ForumId);
+                var forumSlug = _forumService.GetForumSeName(forum);
 
-                _forumService.DeletePost(forumPost);
-
-                //get topic one more time because it can be deleted (first or only post deleted)
-                forumTopic = _forumService.GetTopicById(forumPost.TopicId);
                 if (forumTopic == null)
                 {
                     return Json(new
                     {
-                        redirect = Url.RouteUrl("ForumSlug", new { id = forumId, slug = forumSlug }),
+                        redirect = Url.RouteUrl("ForumSlug", new { id = forum.Id, slug = forumSlug }),
                     });
-                }
+                }                
+
+                _forumService.DeletePost(forumPost);
+                                
                 return Json(new
                 {
                     redirect = Url.RouteUrl("TopicSlug", new { id = forumTopic.Id, slug = _forumService.GetTopicSeName(forumTopic) }),
@@ -681,7 +685,6 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
         [ValidateCaptcha]
         public virtual IActionResult PostCreate(EditForumPostModel model, bool captchaValid)
         {
@@ -756,13 +759,13 @@ namespace Nop.Web.Controllers
 
                     var pageSize = _forumSettings.PostsPageSize > 0 ? _forumSettings.PostsPageSize : 10;
 
-                    var pageIndex = (_forumService.CalculateTopicPageIndex(forumPost.TopicId, pageSize, forumPost.Id) + 1);
+                    var pageIndex = _forumService.CalculateTopicPageIndex(forumPost.TopicId, pageSize, forumPost.Id) + 1;
                     var url = string.Empty;
                     if (pageIndex > 1)
-                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic), pageNumber = pageIndex });
+                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic), pageNumber = pageIndex });
                     else
-                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic) });
-                    return Redirect($"{url}#{forumPost.Id}");
+                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic) });
+                    return LocalRedirect($"{url}#{forumPost.Id}");
                 }
                 catch (Exception ex)
                 {
@@ -793,7 +796,6 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
         [ValidateCaptcha]
         public virtual IActionResult PostEdit(EditForumPostModel model, bool captchaValid)
         {
@@ -807,11 +809,11 @@ namespace Nop.Web.Controllers
             if (!_forumService.IsCustomerAllowedToEditPost(_workContext.CurrentCustomer, forumPost))
                 return Challenge();
 
-            var forumTopic = forumPost.ForumTopic;
+            var forumTopic = _forumService.GetTopicById(forumPost.TopicId);
             if (forumTopic == null)
                 return RedirectToRoute("Boards");
 
-            var forum = forumTopic.Forum;
+            var forum = _forumService.GetForumById(forumTopic.ForumId);
             if (forum == null)
                 return RedirectToRoute("Boards");
 
@@ -871,13 +873,13 @@ namespace Nop.Web.Controllers
                     var url = string.Empty;
                     if (pageIndex > 1)
                     {
-                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic), pageNumber = pageIndex });
+                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic), pageNumber = pageIndex });
                     }
                     else
                     {
-                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic) });
+                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic) });
                     }
-                    return Redirect($"{url}#{forumPost.Id}");
+                    return LocalRedirect($"{url}#{forumPost.Id}");
                 }
                 catch (Exception ex)
                 {
@@ -911,6 +913,7 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost, ActionName("CustomerForumSubscriptions")]
+        [IgnoreAntiforgeryToken]
         public virtual IActionResult CustomerForumSubscriptionsPOST(IFormCollection formCollection)
         {
             foreach (var key in formCollection.Keys)
@@ -944,7 +947,7 @@ namespace Nop.Web.Controllers
             if (forumPost == null)
                 return new NullJsonResult();
 
-            if (!_workContext.CurrentCustomer.IsRegistered())
+            if (!_customerService.IsRegistered(_workContext.CurrentCustomer))
                 return Json(new
                 {
                     Error = _localizationService.GetResource("Forum.Votes.Login"),
